fix bug and enable floating point calculations

+++ src/main/java/org/javaculator/antlr4/handlers/RootAddExprHandler.java
@@ -3,6 +3,7 @@ package org.javaculator.antlr4.handlers;
 import org.javaculator.antlr4.CalcParser;
 import org.javaculator.antlr4.handlers.interfaces.IVisitorExprHandler;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.function.Function;
 
@@ -10,7 +11,7 @@ public class RootAddExprHandler implements IVisitorExprHandler<CalcParser.RootAd
     public static final RootAddExprHandler INSTANCE = new RootAddExprHandler();
 
     @Override
-    public Optional<Integer> handle(CalcParser.RootAddExprContext ctx, Function<CalcParser.AdditiveExprContext, Integer> visitor) {
+    public Optional<BigDecimal> handle(CalcParser.RootAddExprContext ctx, Function<CalcParser.AdditiveExprContext, BigDecimal> visitor) {
         return Optional.ofNullable(ctx.additiveExpr()).map(visitor);
     }
 }
